update performance episode, fixing/reorganising a bit

+++ content/example/integrate.py
@@ -1,6 +1,8 @@
+import numpy as np
+import pandas as pd
+
 def f(x):
     return x ** 2 - x
-
 
 def integrate_f(a, b, N):
     s = 0
@@ -9,11 +11,9 @@ def integrate_f(a, b, N):
         s += f(a + i * dx)
     return s * dx
 
-
 def apply_integrate_f(col_a, col_b, col_N):
     n = len(col_N)
-    res = np.empty(n)
+    res = np.empty(n,dtype=np.float64)
     for i in range(n):
         res[i] = integrate_f(col_a[i], col_b[i], col_N[i])
-
     return res
